Initial AArch64 port in an illumos-y build environment, etc.
Portions contributed by: Michael van der Westhuizen <r1mikey@gmail.com>
Portions contributed by: Jonathan Perkin <jonathan@perkin.org.uk>
Portions contributed by: Richard Lowe <richlowe@richlowe.net>

+++ usr/src/cmd/acctadm/main.c
@@ -137,7 +137,7 @@ main(int argc, char *argv[])
 	(void) setpname(argv[0]);
 
 	for (; optind < argc; optind++) {
-		while ((c = getopt(argc, argv, OPTS)) != (int)EOF) {
+		while ((c = getopt(argc, argv, OPTS)) != EOF) {
 			switch (c) {
 			case 'd':
 				disabled = optarg;
