+++ usr/src/cmd/fs.d/ufs/edquota/edquota.c
@@ -770,7 +770,7 @@ sigsetmask(uint_t omask)
 
 	for (i = 0; i < 32; i++)
 		if (omask & (1 << i)) {
-			if (sigignore(1 << i) == (int)SIG_ERR) {
+			if (sigignore(1 << i) == (int)(intptr_t)SIG_ERR) {
 				(void) fprintf(stderr,
 				    "Bad signal 0x%x\n", (1 << i));
 				exit(31+1);
@@ -787,7 +787,7 @@ sigblock(uint_t omask)
 
 	for (i = 0; i < 32; i++)
 		if (omask & (1 << i)) {
-			if ((temp = sigignore(1 << i)) == (int)SIG_ERR) {
+			if ((temp = sigignore(1 << i)) == (int)(intptr_t)SIG_ERR) {
 				(void) fprintf(stderr,
 				    "Bad signal 0x%x\n", (1 << i));
 				exit(31+1);
