+++ usr/src/cmd/ldapcachemgr/cachemgr_getldap.c
@@ -645,7 +645,7 @@ getldap_get_rootDSE(void *arg)
 	}
 	(void) mutex_unlock(&sig_mutex);
 
-	thr_exit((void *) exitrc);
+	thr_exit((void *)(intptr_t)exitrc);
 
 	return ((void *) NULL);
 }
@@ -840,7 +840,7 @@ getldap_set_serverInfo(server_info_t *head, int reset_bindtime, info_op_t op)
 
 	for (i = 0; i < num_threads; i++) {
 		if (thr_join(tid[i], NULL, &status) == 0) {
-			if ((int)status == NS_LDAP_SUCCESS)
+			if ((int)(intptr_t)status == NS_LDAP_SUCCESS)
 				atleast1 = 1;
 		}
 	}
