+++ usr/src/cmd/svc/configd/configd.h
@@ -707,7 +707,7 @@ int rc_node_delete(rc_node_ptr_t *);
 int rc_node_next_snaplevel(rc_node_ptr_t *, rc_node_ptr_t *);
 
 int rc_node_setup_iter(rc_node_ptr_t *, rc_node_iter_t **, uint32_t,
-    size_t, const char *);
+    uint32_t, const char *);
 
 int rc_iter_next(rc_node_iter_t *, rc_node_ptr_t *, uint32_t);
 int rc_iter_next_value(rc_node_iter_t *, struct rep_protocol_value_response *,
