+++ usr/src/lib/libpkg/common/pkgserv.c
@@ -737,7 +737,7 @@ pkggetentry_named(PKGserver server, const char *path, int *len, int *pathlen)
 	int plen = strlen(path);
 	pkgfilter_t *pcmd = alloca(sizeof (*pcmd) + plen);
 	char *result;
-	unsigned int rlen;
+	size_t rlen;
 
 	pcmd->cmd = PKG_FINDFILE;
 	*pathlen = pcmd->len = plen;
